Move polygon drawing classes to datacollection package
