+++ ground/src/main/java/com/google/android/ground/ui/datacollection/PolygonDrawingTaskFragment.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.home.mapcontainer
+package com.google.android.ground.ui.datacollection
 
 import android.os.Bundle
 import android.view.LayoutInflater
