+++ ground/src/main/res/layout/polygon_drawing_task_frag.xml
@@ -22,7 +22,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.home.mapcontainer.PolygonDrawingViewModel" />
+      type="com.google.android.ground.ui.datacollection.PolygonDrawingViewModel" />
     <import type="android.view.View" />
   </data>
